2024.04.29 - 내가 찜한 사람 목록

+++ src/main/java/com/example/backend/domain/User.java
@@ -75,7 +75,7 @@ public class User {
     @ElementCollection
     @CollectionTable(name = "userLike", joinColumns = @JoinColumn(name = "user_id"))
     @Column(name = "favorite_id")
-    private Set<Integer> userLike = new HashSet<>();
+    private Set<Long> userLike = new HashSet<>();
 
     @Builder.Default
     @OneToMany(mappedBy = "user")
